refactor: 레거시 클래스를 신규 클래스로 대체
