+++ backend/src/main/java/com/woowacourse/kkogkkog/application/CouponService.java
@@ -1,6 +1,7 @@
 package com.woowacourse.kkogkkog.application;
 
 import com.woowacourse.kkogkkog.application.dto.CouponResponse;
+import com.woowacourse.kkogkkog.application.dto.CouponSaveRequest;
 import com.woowacourse.kkogkkog.domain.Coupon;
 import com.woowacourse.kkogkkog.domain.CouponStatus;
 import com.woowacourse.kkogkkog.domain.CouponType;
@@ -9,7 +10,8 @@ import com.woowacourse.kkogkkog.domain.repository.CouponRepository;
 import com.woowacourse.kkogkkog.domain.repository.MemberRepository;
 import com.woowacourse.kkogkkog.exception.coupon.CouponNotFoundException;
 import com.woowacourse.kkogkkog.exception.member.MemberNotFoundException;
-import com.woowacourse.kkogkkog.presentation.dto.CouponCreateRequest;
+import java.util.List;
+import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -25,35 +27,45 @@ public class CouponService {
         this.memberRepository = memberRepository;
     }
 
-    public Long save(CouponCreateRequest couponCreateRequest) {
-        Coupon coupon = saveCoupon(couponCreateRequest);
-        Coupon savedCoupon = couponRepository.save(coupon);
-
-        return savedCoupon.getId();
-    }
-
-    private Coupon saveCoupon(CouponCreateRequest couponCreateRequest) {
-        Member sender = findMember(couponCreateRequest.getSenderId());
-        Member receiver = findMember(couponCreateRequest.getReceiverId());
-        String modifier = couponCreateRequest.getModifier();
-        String message = couponCreateRequest.getMessage();
-        String backgroundColor = couponCreateRequest.getBackgroundColor();
-        CouponType couponType = CouponType.of(couponCreateRequest.getCouponType());
-        CouponStatus couponStatus = CouponStatus.READY;
-
-        return new Coupon(sender, receiver, modifier, message, backgroundColor, couponType, couponStatus);
-    }
-
-    private Member findMember(Long memberId) {
-        return memberRepository.findById(memberId)
-                .orElseThrow(MemberNotFoundException::new);
-    }
-
     @Transactional(readOnly = true)
     public CouponResponse findById(Long couponId) {
         Coupon coupon = couponRepository.findById(couponId)
                 .orElseThrow(CouponNotFoundException::new);
 
         return CouponResponse.of(coupon);
+    }
+
+    public List<CouponResponse> save(CouponSaveRequest couponSaveRequest) {
+        List<Coupon> coupons = toCoupons(couponSaveRequest);
+        List<Coupon> savedCoupons = couponRepository.saveAll(coupons);
+        return savedCoupons.stream()
+                .map(CouponResponse::of)
+                .collect(Collectors.toList());
+    }
+
+    private List<Coupon> toCoupons(CouponSaveRequest couponSaveRequest) {
+        Member sender = findMember(couponSaveRequest.getSenderId());
+        List<Member> receivers = findMembers(couponSaveRequest.getReceivers());
+        String modifier = couponSaveRequest.getModifier();
+        String message = couponSaveRequest.getMessage();
+        String backgroundColor = couponSaveRequest.getBackgroundColor();
+        CouponType couponType = CouponType.valueOf(couponSaveRequest.getCouponType());
+        CouponStatus couponStatus = CouponStatus.READY;
+        return receivers.stream()
+                .map(it -> new Coupon(sender, it, modifier, message, backgroundColor, couponType, couponStatus))
+                .collect(Collectors.toList());
+    }
+
+    private List<Member> findMembers(List<Long> memberIds) {
+        List<Member> receivers = memberRepository.findAllById(memberIds);
+        if (memberIds.size() != receivers.size()) {
+            throw new MemberNotFoundException();
+        }
+        return receivers;
+    }
+
+    private Member findMember(Long memberId) {
+        return memberRepository.findById(memberId)
+                .orElseThrow(MemberNotFoundException::new);
     }
 }
