+++ backend/src/main/java/com/woowacourse/kkogkkog/application/dto/CouponResponse.java
@@ -10,29 +10,36 @@ import lombok.NoArgsConstructor;
 public class CouponResponse {
 
     private Long id;
-    private String senderName;
-    private String receiverName;
-    private String backgroundColor;
+    private CouponMemberResponse sender;
+    private CouponMemberResponse receiver;
     private String modifier;
     private String message;
+    private String backgroundColor;
     private String couponType;
     private String couponStatus;
 
-    public CouponResponse(Long id, String senderName, String receiverName, String backgroundColor, String modifier,
-                          String message, String couponType, String couponStatus) {
+    public CouponResponse(Long id, CouponMemberResponse sender,
+                          CouponMemberResponse receiver, String modifier, String message,
+                          String backgroundColor, String couponType, String couponStatus) {
         this.id = id;
-        this.senderName = senderName;
-        this.receiverName = receiverName;
-        this.backgroundColor = backgroundColor;
+        this.sender = sender;
+        this.receiver = receiver;
         this.modifier = modifier;
         this.message = message;
+        this.backgroundColor = backgroundColor;
         this.couponType = couponType;
         this.couponStatus = couponStatus;
     }
 
     public static CouponResponse of(Coupon coupon) {
-        return new CouponResponse(coupon.getId(), coupon.getSender().getNickname(), coupon.getReceiver().getNickname(),
-                coupon.getBackgroundColor(), coupon.getModifier(), coupon.getMessage(),
-                coupon.getCouponType().getValue(), coupon.getCouponStatus().name());
+        return new CouponResponse(
+                coupon.getId(),
+                CouponMemberResponse.of(coupon.getSender()),
+                CouponMemberResponse.of(coupon.getReceiver()),
+                coupon.getModifier(),
+                coupon.getMessage(),
+                coupon.getBackgroundColor(),
+                coupon.getCouponType().name(),
+                coupon.getCouponStatus().name());
     }
 }
