+++ backend/src/main/java/com/woowacourse/kkogkkog/presentation/CouponController.java
@@ -1,7 +1,7 @@
 package com.woowacourse.kkogkkog.presentation;
 
-import com.woowacourse.kkogkkog.application.CouponService2;
-import com.woowacourse.kkogkkog.application.dto.CouponResponse2;
+import com.woowacourse.kkogkkog.application.CouponService;
+import com.woowacourse.kkogkkog.application.dto.CouponResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,15 +12,15 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/coupons")
 public class CouponController {
 
-    private final CouponService2 couponService;
+    private final CouponService couponService;
 
-    public CouponController(CouponService2 couponService) {
+    public CouponController(CouponService couponService) {
         this.couponService = couponService;
     }
 
     @GetMapping("/{couponId}")
-    public ResponseEntity<CouponResponse2> showCoupon(@PathVariable Long couponId) {
-        CouponResponse2 couponResponse = couponService.findById(couponId);
+    public ResponseEntity<CouponResponse> showCoupon(@PathVariable Long couponId) {
+        CouponResponse couponResponse = couponService.findById(couponId);
         return ResponseEntity.ok(couponResponse);
     }
 }
