+++ backend/src/test/java/com/woowacourse/kkogkkog/application/CouponServiceTest.java
@@ -1,27 +1,30 @@
 package com.woowacourse.kkogkkog.application;
 
+import static com.woowacourse.kkogkkog.fixture.MemberFixture.ARTHUR;
+import static com.woowacourse.kkogkkog.fixture.MemberFixture.JEONG;
+import static com.woowacourse.kkogkkog.fixture.MemberFixture.LEO;
+import static com.woowacourse.kkogkkog.fixture.MemberFixture.NON_EXISTING_MEMBER;
+import static com.woowacourse.kkogkkog.fixture.MemberFixture.ROOKIE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.springframework.test.annotation.DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD;
 
 import com.woowacourse.kkogkkog.application.dto.CouponResponse;
-import com.woowacourse.kkogkkog.domain.CouponStatus;
+import com.woowacourse.kkogkkog.application.dto.CouponSaveRequest;
+import com.woowacourse.kkogkkog.domain.Member;
 import com.woowacourse.kkogkkog.domain.repository.MemberRepository;
 import com.woowacourse.kkogkkog.exception.coupon.CouponNotFoundException;
-import com.woowacourse.kkogkkog.fixture.MemberFixture;
-import com.woowacourse.kkogkkog.presentation.dto.CouponCreateRequest;
+import com.woowacourse.kkogkkog.exception.member.MemberNotFoundException;
+import java.util.List;
+import java.util.stream.Collectors;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.transaction.annotation.Transactional;
 
-@SpringBootTest
 @Transactional
-@DirtiesContext(classMode = BEFORE_EACH_TEST_METHOD)
-public class CouponServiceTest {
+public class CouponServiceTest extends ServiceTest {
 
     @Autowired
     private CouponService couponService;
@@ -29,38 +32,75 @@ public class CouponServiceTest {
     @Autowired
     private MemberRepository memberRepository;
 
+    @Override
     @BeforeEach
     void setUp() {
-        memberRepository.save(MemberFixture.ROOKIE);
-        memberRepository.save(MemberFixture.ARTHUR);
-        memberRepository.save(MemberFixture.JEONG);
+        super.setUp();
+        memberRepository.save(JEONG);
+        memberRepository.save(LEO);
+        memberRepository.save(ROOKIE);
+        memberRepository.save(ARTHUR);
     }
 
-    @Test
-    @DisplayName("쿠폰을 생성할 수 있다.")
-    void save() {
-        CouponCreateRequest couponCreateRequest = new CouponCreateRequest(1L, 2L, "red", "한턱내는", "추가 메세지", "커피");
-        Long couponId = couponService.save(couponCreateRequest);
-        assertThat(couponId).isNotNull();
+    @DisplayName("단일 쿠폰을 조회할 수 있다.")
+    @Nested
+    class FindByIdTest {
+
+        @DisplayName("존재하는 쿠폰을 조회하는 경우 성공한다.")
+        @Test
+        void findById() {
+            List<CouponResponse> savedCoupons = couponService.save(toCouponSaveRequest(JEONG, List.of(LEO, ROOKIE)));
+
+            CouponResponse expected = savedCoupons.get(0);
+            CouponResponse actual = couponService.findById(expected.getId());
+
+            assertThat(actual).usingRecursiveComparison().isEqualTo(expected);
+        }
+
+        @Test
+        @DisplayName("존재하지 않는 쿠폰을 조회할 경우 예외가 발생한다.")
+        void findById_notFound() {
+            assertThatThrownBy(() -> couponService.findById(1L))
+                    .isInstanceOf(CouponNotFoundException.class);
+        }
     }
 
-    @Test
-    @DisplayName("쿠폰을 조회할 수 있다.")
-    void findById() {
-        CouponCreateRequest couponCreateRequest = new CouponCreateRequest(1L, 2L, "red", "한턱내는", "추가 메세지", "커피");
-        Long couponId = couponService.save(couponCreateRequest);
+    @DisplayName("복수의 쿠폰을 저장할 수 있다")
+    @Nested
+    class SaveTest {
 
-        CouponResponse couponResponse = couponService.findById(couponId);
+        @Test
+        @DisplayName("받는 사람으로 지정한 사용자들에게 동일한 내용의 쿠폰이 발급된다.")
+        void save() {
+            CouponSaveRequest couponSaveRequest = toCouponSaveRequest(ROOKIE, List.of(ARTHUR, JEONG, LEO));
+            List<CouponResponse> createdCoupons = couponService.save(couponSaveRequest);
 
-        assertThat(couponResponse).usingRecursiveComparison()
-                .isEqualTo(new CouponResponse(couponId, "루키", "아서", "red", "한턱내는", "추가 메세지", "커피",
-                        CouponStatus.READY.name()));
+            assertThat(createdCoupons.size()).isEqualTo(3);
+        }
+
+        @Test
+        @DisplayName("존재하지 않는 사용자가 쿠폰을 보내려는 경우 예외가 발생한다.")
+        void save_senderNotFound() {
+            CouponSaveRequest couponSaveRequest = toCouponSaveRequest(NON_EXISTING_MEMBER, List.of(ARTHUR, LEO));
+
+            assertThatThrownBy(() -> couponService.save(couponSaveRequest))
+                    .isInstanceOf(MemberNotFoundException.class);
+        }
+
+        @Test
+        @DisplayName("존재하지 않는 사용자에게 쿠폰을 보내려는 경우 예외가 발생한다.")
+        void save_receiverNotFound() {
+            CouponSaveRequest couponSaveRequest = toCouponSaveRequest(JEONG, List.of(ARTHUR, NON_EXISTING_MEMBER));
+
+            assertThatThrownBy(() -> couponService.save(couponSaveRequest))
+                    .isInstanceOf(MemberNotFoundException.class);
+        }
     }
 
-    @Test
-    @DisplayName("존재하지 않는 쿠폰을 조회할 경우 예외가 발생한다")
-    void findById_notFound() {
-        assertThatThrownBy(() -> couponService.findById(1L))
-                .isInstanceOf(CouponNotFoundException.class);
+    private CouponSaveRequest toCouponSaveRequest(Member sender, List<Member> receivers) {
+        List<Long> receiverIds = receivers.stream()
+                .map(Member::getId)
+                .collect(Collectors.toList());
+        return new CouponSaveRequest(sender.getId(), receiverIds, "#123456", "한턱내는", "추가 메세지", "COFFEE");
     }
 }
